+++ src/routes/map/capacity/+page.svelte
@@ -10,6 +10,7 @@
 	import ErrorMessage from '$components/ErrorMessage.svelte';
 	import WarningMessage from '$components/WarningMessage.svelte';
 	import type { ContextMenuItem } from '$components/ContextMenu.svelte';
+	import Slider from '$components/forms/Slider.svelte';
 
 	import type { ActivatedSolution, System } from '$lib/types';
 	import { fetchTotal } from '$lib/temple';
@@ -45,7 +46,7 @@
 	let selectedTechnologyType: TechnologyType = $state('conversion');
 	let selectedStorageType: StorageType = $state('energy');
 	let selectedCarrier: string | null = $state(null);
-	let selectedYear: string | null = $state(null);
+	let selectedYear: number = $state(0);
 	let selectedMap: string | null = $state('world-3');
 
 	let solutionLoading: boolean = $state(false);
@@ -123,7 +124,7 @@
 					[QUERY_PARAM_KEYS.solution]: selectedSolution.solution_name,
 					[QUERY_PARAM_KEYS.scenario]: selectedSolution.scenario_name,
 					[QUERY_PARAM_KEYS.carrier]: selectedCarrier,
-					[QUERY_PARAM_KEYS.year]: selectedYear,
+					[QUERY_PARAM_KEYS.year]: selectedYear.toString(),
 					[QUERY_PARAM_KEYS.node]: pie.label
 				})
 			});
@@ -224,14 +225,14 @@
 			</FilterSection>
 			{#if selectedSolution !== null && selectedCarrier !== null}
 				<FilterSection title="Data Selection">
-					<Dropdown
+					<Slider
 						bind:value={selectedYear}
-						options={years.map((year) => year.toString())}
+						min={years[0]}
+						max={years[years.length - 1]}
+						step={selectedSolution.detail.system.interval_between_years}
 						label="Year"
 						urlParam={QUERY_PARAM_KEYS.year}
-						unsetIfInvalid
-						default={years.length > 0 ? years[0].toString() : null}
-					></Dropdown>
+					></Slider>
 					<Dropdown bind:value={selectedMap} options={maps} label="Map"></Dropdown>
 				</FilterSection>
 			{/if}